% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataprep.R
\name{expandlevels}
\alias{expandlevels}
\title{expand all levels for factor columns}
\usage{
expandlevels(dataframe, factorcols)
}
\arguments{
\item{dataframe}{a data frame}

\item{factorcols}{a list of factor columns to expand
if it's not provided, will use all factor columns}
}
\value{
a dataframe with factor columns expanded
}
\description{
expand all levels for factor columns
}

