% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataprep.R
\name{removefactors}
\alias{removefactors}
\title{remove factors from a data frame}
\usage{
removefactors(dataframe, nlevels = 50)
}
\arguments{
\item{dataframe}{a data frame}

\item{nlevels}{an integer - number of max levels of factor columns allowed}
}
\value{
a data frame whose factor columns with either > nlevels
of levels or one level are removed
}
\description{
remove factors from a data frame
}

