% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataprep.R
\name{splitdf}
\alias{splitdf}
\title{split a data frame into training and testing sets}
\usage{
splitdf(dataframe, seed = 10001, ratio = 2)
}
\arguments{
\item{dataframe}{}

\item{seed}{(optional) - seeding the split}

\item{ratio}{(optional) - 1/ratio of total set is designated to training}
}
\value{
a list - trainset and testset
}
\description{
split a data frame into training and testing sets
}

