% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotcormatrix.R
\name{plotcormatrix}
\alias{plotcormatrix}
\title{plot correlation matrix}
\usage{
plotcormatrix(dataframe, roundn = 2, type = "lower", corfix = FALSE)
}
\arguments{
\item{dataframe}{}

\item{roundn}{(optional) - an integer number of decimal places to round correlation coefficients, default = 2}

\item{type}{(optional) - 'lower' (default) for lower triangular matrix, 'upper' for upper triangular matrix, and 'full' for full matrix}

\item{corfix}{(optiona) - 'FALSE' (default) for natural output, 'TRUE' for square output}
}
\value{
correlation matrix
}
\description{
plot correlation matrix
}

